## Summary
- Fix MXFP4 fallback on gfx906: keep scales in `uint8` on the right device, cast router logits, and dequant via `fused_experts`, so MXFP4 checkpoints run in FP16 emulation without Quark errors (requires `pip install amd-quark` and `apt-get install -y libxml2` in the current `nlzy/vllm-gfx906` Docker image).
- Make the GPT-OSS MoE router pick a dtype supported on ROCm and cast inputs/outputs to avoid BF16/Half mismatches.
- Add `VLLM_DISABLE_GPTOSS_HARMONY` so GPT-OSS reasoning wrappers can be toggled; the config no longer forces the parser when the flag is set, and both OpenAI front-ends skip Harmony in that case.
- Update `benchmark_moe.py` to honor `ROCR_VISIBLE_DEVICES`/`CUDA_VISIBLE_DEVICES`, pin Ray workers to `cuda:0`, and guard kernel launches with the correct device context.

## Testing
- `benchmarks/kernels/benchmark_moe.py`  
  - Tuned `Qwen3-30B-A3B-Thinking-2507-GPTQ-Int8` with `--enable-expert-parallel --tensor-parallel-size 4`; generated `E=32,N=768,device_name=Radeon_Instinct_MI50_32GB,dtype=int8_w8a16.json`.
- vLLM serve on MI50 (gfx906)  
  - MXFP4 checkpoints load and infer in FP16 emulation mode (after installing `amd-quark` and `libxml2`).  
  - `VLLM_DISABLE_GPTOSS_HARMONY=0/1` flips GPT-OSS responses between reasoning-wrapped and plain text (this adjustment is for some benchmarks script).
- Performance: at ~5120-token prompts, tuning improves prefill throughput from 1963 tok/s → 2326 tok/s and decode throughput from 39 tok/s → 47 tok/s (~20% uplift).