@Said-Akbar 2025-09-21
Hi @tomylin890 ,

Thank you for creating this PR to support gpt-oss. Can you please share if you used 2xMI50 32GB or 4x for your tests?
Also, when tuning, did you use batch_size=1 ? Thanks!