@tomylin890 2025-09-21
> Can you please share your tuned output file? This file: E=32,N=768,device_name=Radeon_Instinct_MI50_32GB,dtype=int8_w8a16.json

Here is json config I generated on my platform (5700X + 4xMI50 under PEX88096 PCIe Switch, VBIOS 113-D1631700-111, PCIe P2P is enabled)

Model: https://huggingface.co/QuantTrio/Qwen3-30B-A3B-Thinking-2507-GPTQ-Int8

```
{
    "1": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "2": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "4": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "8": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 64,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "16": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "24": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 8,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "32": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 32,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 8,
        "num_warps": 4,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "48": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 128,
        "BLOCK_SIZE_K": 64,
        "GROUP_SIZE_M": 1,
        "num_warps": 2,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "64": {
        "BLOCK_SIZE_M": 16,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 64,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "96": {
        "BLOCK_SIZE_M": 32,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "128": {
        "BLOCK_SIZE_M": 32,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "256": {
        "BLOCK_SIZE_M": 32,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "512": {
        "BLOCK_SIZE_M": 32,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 8,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "1024": {
        "BLOCK_SIZE_M": 32,
        "BLOCK_SIZE_N": 64,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 8,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "1536": {
        "BLOCK_SIZE_M": 64,
        "BLOCK_SIZE_N": 32,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 8,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "2048": {
        "BLOCK_SIZE_M": 128,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 4,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "3072": {
        "BLOCK_SIZE_M": 128,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 1,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    },
    "4096": {
        "BLOCK_SIZE_M": 128,
        "BLOCK_SIZE_N": 16,
        "BLOCK_SIZE_K": 32,
        "GROUP_SIZE_M": 4,
        "num_warps": 1,
        "num_stages": 2,
        "waves_per_eu": 0
    }
}
```

Some News: I also make Qwen3Next work on nlzy's repo. But still optimizing, there is some performance test:

comment: Qwen3-Next-80B-A3B-Thinking-AWQ-4bit 4xMI50 without json config (not generated)

| prompt length (tokens) | prefill time (ms) | prefill speed (tokens/s) | output length (tokens) | output time (ms) | output speed (tokens/s) |
|---------------------|-------------------|-------------------------|-------------------|---------------|---------------------|
| 512 | 542.00 | 944.65 | 128 | 5501.00 | 23.27 |
| 5632 | 4460.00 | 1262.78 | 128 | 6125.00 | 20.90 |
| 10752 | 10364.00 | 1037.44 | 128 | 6765.00 | 18.92 |
| 15872 | 17625.00 | 900.54 | 128 | 7414.00 | 17.26 |
| 20992 | 27014.00 | 777.08 | 128 | 8016.00 | 15.97 |
| 26112 | 37575.00 | 694.93 | 128 | 9243.00 | 13.85 |
| 31232 | 50489.00 | 618.59 | 128 | 9898.00 | 12.93 |
| 36352 | 64395.00 | 564.52 | 128 | 10501.00 | 12.19 |
| 41472 | 80696.00 | 513.93 | 128 | 11120.00 | 11.51 |

 I think it also needs to be optimized too.., It is a big work because I choose backport instead of update to upstream vllm v0.10.2.